reorganize file structure for clearer app components
