+++ app/_lib/ai.ts
