+++ app/api/auth/[...nextauth]/route.ts
@@ -2,7 +2,7 @@ import NextAuth from "next-auth"
 import GoogleProvider from "next-auth/providers/google"
 import { PrismaAdapter } from "@auth/prisma-adapter"
 
-import { prisma } from "@/lib/prisma"
+import { prisma } from "@/app/_lib/prisma"
 
 // Configure NextAuth with v4 syntax
 const handler = NextAuth({
